Refactor tasker controller and service, add getListOfTaskers endpoint
This commit refactors the tasker controller and service by adding a new endpoint called "getListOfTaskers". This endpoint allows users to retrieve a list of taskers based on their location and the services they provide. The endpoint accepts query parameters for longitude, latitude, services, and maxDistance.

The tasker controller now includes a new method called "getListOfTaskers" which calls the corresponding method in the tasker service. The tasker service method validates the query parameters and checks if the specified services exist in the database. It then calls the "listTaskers" method in the tasker DAO to retrieve the taskers based on the provided parameters.

This commit also includes some minor changes to the tasker DAO and tasker model files to support the new functionality.