Add image cover field and improve image upload
Added an image cover field to the task model and interface. This field contains a URL and a publicId. Also, refactored the image upload middleware to support both single and multiple file uploads. The upload functions now take in parameters to specify the file name and the maximum count of files to be uploaded.

Updated the service and user routes to use the new upload functions from the middleware. The service image and profile image upload now use the uploadSingleFile function from the middleware.

This change improves the flexibility of the image upload feature and allows for more detailed image information to be stored in the task model.