Refactor image upload middleware
This commit refactors the image upload middleware in the application. The multer options function has been moved to the bottom of the file for better readability. The filename generation logic has been updated to include 'cover' in the filename if the file fieldname is 'imageCover'. This makes it easier to identify cover images in the uploads directory. The rest of the changes are minor code cleanups and reordering for better readability and maintainability.