Refactor image handling in service model
This commit refactors the way images are handled in the service model. The image field has been changed from a simple string to an object that includes a URL and a publicId. This allows for more robust handling of images, including the ability to track where they are stored.

In addition, the uploadServiceImage method in the service service has been updated to use the new image field structure. The method now uploads the image directly from the file buffer, rather than saving it to the server first. This simplifies the process and removes the need to manually delete the file from the server after it has been uploaded.

Finally, the service route has been updated to include a check for a valid MongoDB ID before uploading an image. This ensures that the ID provided is valid before attempting to upload an image.