Update API details and improve auth middleware
Updated the API details in the swagger.json file to reflect the new application name and URL. The contact information, server URL, and authorization URL have been updated to match the new application details.

In the auth middleware, added a comment to clarify the purpose of the 'iat' variable. Also added a TODO comment to suggest moving the password change check to the user model for better code organization.