Refactor checkout process and add phone number to User
This commit refactors the checkout process in the offer service to accept the entire user object instead of just the user ID. It also adds a phone number field to the User interface and updates the Paymob service to use the user's phone number when generating payment tokens. Additionally, the Paymob service now has an authenticate method to retrieve the Paymob token, and the createOrder method no longer hardcodes items but expects an empty array. Error handling is improved in the generatePaymentToken method, and the initiateWalletPayment method now requires a phone number.